src/backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "buildersmty-backend"
version = "0.1.0"
description = "Backend for Builders MTY web and discord"
authors = [
    {name = "Raul R. Gonzalez", email = "raul@buildersmty.com.mx"}
]

requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.2",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.23",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "discord-py>=2.3.2",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.2",
    "lxml>=5.1.0",
    "python-multipart>=0.0.20",
    "email-validator>=2.1.0.post1",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "supabase>=2.0.0",
    "openai>=1.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]